+++ src/algo/predecessor.rs
@@ -1,7 +1,7 @@
 //! Predecessor trees
 //!
 //! Predecessor trees are produced by many graph algorithms, e.g., those defined
-//! in [`crate::algo::bfs`] and [`crate::algo::dijkstra`].
+//! in [`bfs`](crate::algo::bfs) and [`dijkstra`](crate::algo::dijkstra).
 
 use std::collections::HashSet;