+++ src/gen/mod.rs
@@ -2,6 +2,10 @@
 //!
 //! Generate graphs with different properties.
 
+pub mod empty;
 pub mod linear;
 
-pub use linear::Linear;
+pub use {
+    empty::Empty,
+    linear::Linear,
+};