+++ src/op/add_edge.rs
@@ -66,13 +66,14 @@ use {
 ///
 /// ## `AddEdge` and `RemoveEdge`
 ///
-/// Types that also implement [`crate::op::RemoveEdge`] should ensure that
-/// [`crate::op::prop::add_edge_remove_edge`] holds.
+/// Types that also implement [`RemoveEdge`](crate::op::RemoveEdge) should
+/// ensure that [`add_edge_remove_edge`](crate::op::prop::add_edge_remove_edge)
+/// holds.
 ///
 /// ## `AddEdge` and `IsEdge`
 ///
-/// Types that also implement [`crate::op::IsEdge`] should ensure that
-/// [`crate::op::prop::add_edge_is_edge`] holds.
+/// Types that also implement [`IsEdge`](crate::op::IsEdge) should ensure that
+/// [`add_edge_is_edge`](crate::op::prop::add_edge_is_edge) holds.
 pub trait AddEdge {
     /// Add an edge from `s` to `t`.
     ///