+++ src/op/add_weighted_edge.rs
@@ -69,13 +69,15 @@ use {
 ///
 /// ## `AddWeightedEdge` and `RemoveEdge`
 ///
-/// Types that also implement [`crate::op::RemoveEdge`] should ensure that
-/// [`crate::op::prop::add_weighted_edge_remove_edge`] holds.
+/// Types that also implement [`RemoveEdge`](crate::op::RemoveEdge) should
+/// ensure that
+/// [`add_weighted_edge_remove_edge`](crate::op::prop::add_weighted_edge_remove_edge) holds.
 ///
 /// ## `AddWeightedEdge` and `IsEdge`
 ///
-/// Types that also implement [`crate::op::IsEdge`] should ensure that
-/// [`crate::op::prop::add_weighted_edge_is_edge`] holds.
+/// Types that also implement [`IsEdge`](crate::op::IsEdge) should ensure that
+/// [`add_weighted_edge_is_edge`](crate::op::prop::add_weighted_edge_is_edge)
+/// holds.
 pub trait AddWeightedEdge<W> {
     /// Add an edge from `s` to `t` with weight `w`.
     ///