+++ src/op/remove_edge.rs
@@ -121,18 +121,20 @@ use {
 ///
 /// ## `RemoveEdge` and `AddEdge`
 ///
-/// Types that also implement [`crate::op::AddEdge`] should ensure that
-/// [`crate::op::prop::add_edge_remove_edge`] holds.
+/// Types that also implement [`AddEdge`](crate::op::AddEdge) should ensure that
+/// [`add_edge_remove_edge`](crate::op::prop::add_edge_remove_edge) holds.
 ///
 /// ## `RemoveEdge` and `AddWeightedEdge`
 ///
-/// Types that also implement [`crate::op::AddWeightedEdge`] should ensure that
-/// [`crate::op::prop::add_weighted_edge_remove_edge`] holds.
+/// Types that also implement [`AddWeightedEdge`](crate::op::AddWeightedEdge)
+/// should ensure that
+/// [`add_weighted_edge_remove_edge`](crate::op::prop::add_weighted_edge_remove_edge)
+/// holds.
 ///
 /// ## `RemoveEdge` and `IsEdge`
 ///
-/// Types that also implement [`crate::op::IsEdge`] should ensure that
-/// [`crate::op::prop::remove_edge_is_edge`] holds.
+/// Types that also implement [`IsEdge`](crate::op::IsEdge) should ensure that
+/// [`remove_edge_is_edge`](crate::op::prop::remove_edge_is_edge) holds.
 pub trait RemoveEdge {
     /// Remove the edge from `s` to `t`. Returns whether the edge was removed.
     ///